backend/requirements.txt
Flask==2.3.3
paho-mqtt==1.6.1
firebase-admin==5.2.0
python-dotenv==1.0.0
Flask-Cors==3.1.1
google-cloud-firestore==2.11.0
